add authorization for Subbreddits

+++ app/Http/Controllers/SubbredditsController.php
@@ -17,9 +17,9 @@ class SubbredditsController extends Controller
      */
     public function index()
     {
-        // pagination returns errors on the AJAX requests
+        // pagination returns errors on the AJAX requests, can use all() tmp instead
         // need to find a fix in the multi-layered JSON selector
-        return Subbreddit::all();
+        return Subbreddit::paginate(15);
     }
 
     /**
@@ -31,7 +31,7 @@ class SubbredditsController extends Controller
     public function store(Request $request)
     {
         $subbreddit = new Subbreddit;
-        $subbreddit->user_id = \Auth::user()->id;
+        $subbreddit->user_id = auth()->user()->id;
         $subbreddit->name = $request->name;
         $subbreddit->description = $request->description;
         $subbreddit->save();
@@ -64,6 +64,7 @@ class SubbredditsController extends Controller
     public function update(Request $request, $id)
     {
         $subbreddit = Subbreddit::findOrFail($id);
+        $this->authorize('update-destroy-subbreddit', $subbreddit);
         $subbreddit->name = $request->name;
         $subbreddit->description = $request->description;
         $subbreddit->save();
@@ -80,6 +81,7 @@ class SubbredditsController extends Controller
     public function destroy($id)
     {
         $subbreddit = Subbreddit::findOrFail($id);
+        $this->authorize('update-destroy-subbreddit', $subbreddit);
         $subbreddit->delete();
         return $subbreddit;
     }
